% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltalk.R
\name{circosPlot}
\alias{circosPlot}
\title{To present a circos plot}
\usage{
circosPlot(Interact, ident = NULL)
}
\arguments{
\item{Interact}{Interact list returned by findLRpairs}

\item{ident}{To highlight the interaction of a specific identity class between others; if 'NULL', plot interaction for all identity classes}
}
\value{
Circos plot showing the ligand-receptor interaction
}
\description{
To present a circos plot
}
