% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltalk.R
\name{dotPlot}
\alias{dotPlot}
\title{To present a dot plot for specific ligand-receptor pairs in specific clusters}
\usage{
dotPlot(all.marker.dat, Interact, ligand.ident = NULL, receptor.ident = NULL)
}
\arguments{
\item{all.marker.dat}{Data frame containing information of marker genes}

\item{Interact}{Interact list returned by findLRpairs}

\item{ligand.ident}{Vector containing the ligand ident}

\item{receptor.ident}{Vector containing the receptor ident}
}
\value{
Dotplot showing the ligand-receptor interaction between the selected ligand.ident and receptor.ident
}
\description{
To present a dot plot for specific ligand-receptor pairs in specific clusters
}
