% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltalk.R
\name{findDEGs}
\alias{findDEGs}
\title{To find marker ligands and marker receptors}
\usage{
findDEGs(expr.mat, lable, method = "wilcox.test", p.adjust = "BH")
}
\arguments{
\item{expr.mat}{Matrix or data frame of expression matrix, with genes in rows and cells in columns}

\item{lable}{Vector of identity classes of cells in the expression matrix}

\item{method}{Method used for differential expression test, either 'wilcox.test' or 't.test'}
}
\value{
Data frame containing the differential expression test
}
\description{
To find marker ligands and marker receptors
}
