% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltalk.R
\name{findLRpath}
\alias{findLRpath}
\title{To find those pathways in which the genesets show overlap with the marker ligand and receptor genes in our dataset}
\usage{
findLRpath(Interact = Interact)
}
\arguments{
\item{Interact}{Interact list returned by findLRpairs}
}
\value{
Interact list containing the ligand-receptor interaction information and the pathways showing overlap with the marker ligand and receptor genes in the dataset
}
\description{
To find those pathways in which the genesets show overlap with the marker ligand and receptor genes in our dataset
}
