% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltalk.R
\name{receptorPathPlot}
\alias{receptorPathPlot}
\title{To present the interact and variable pathways in a line plot, for a specific ident}
\usage{
receptorPathPlot(
  Interact,
  gsva.mat,
  ident.path.dat,
  p.thre = 0.05,
  n = 10,
  order = 1:20,
  ident.size = 1,
  label.dist = 0.2,
  label.size = 1,
  line.size = 1,
  ident.line.alpha = 1,
  ident.line.width = 1,
  path.line.width = 1,
  path.line.alpha = 1,
  path.size = 1
)
}
\arguments{
\item{Interact}{Interact list returned by findLRpairs}

\item{gsva.mat}{pathway score}

\item{ident.path.dat}{ident.path.dat for the selected ident}

\item{p.thre}{Order of identity classes in the plot}

\item{n}{Top n variable pathways}

\item{order}{Whether to hide the isolated identity (those identity class in which cells show no interact with cells in other ientity class) or not; defult is TRUE}

\item{ident.size}{line.width}
}
\value{
String with idents pasted
}
\description{
To present the interact and variable pathways in a line plot, for a specific ident
}
